@@ -1,5 +1,5 @@
 /*
- *   Copyright 2017 Benoit LETONDOR
+ *   Copyright 2019 Benoit LETONDOR
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -16,12 +16,12 @@
 
 package com.benoitletondor.mvp.core.activity;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
 
 import com.benoitletondor.mvp.core.SpyPresenter;
-import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
+import com.benoitletondor.mvp.core.presenter.PresenterFactory;
 import com.benoitletondor.mvp.core.view.View;
 import com.benoitletondor.mvp.core.view.impl.BaseMVPActivity;
 
@@ -32,24 +32,17 @@ import com.benoitletondor.mvp.core.view.impl.BaseMVPActivity;
  */
 public class MVPActivity extends BaseMVPActivity<SpyPresenter<MVPActivity>, MVPActivity> implements View
 {
-    @Override
-    protected PresenterFactory<SpyPresenter<MVPActivity>> getPresenterFactory()
-    {
-        return new PresenterFactory<SpyPresenter<MVPActivity>>()
-        {
-            @NonNull
-            @Override
-            public SpyPresenter<MVPActivity> create()
-            {
-                return new SpyPresenter<>();
-            }
-        };
-    }
-
     @Nullable
     @VisibleForTesting
     public SpyPresenter<MVPActivity> getPresenter()
     {
         return mPresenter;
+    }
+
+    @Override
+    @NonNull
+    protected PresenterFactory<SpyPresenter<MVPActivity>> getPresenterFactory()
+    {
+        return SpyPresenter::new;
     }
 }
