@@ -1,5 +1,5 @@
 /*
- *   Copyright 2017 Benoit LETONDOR
+ *   Copyright 2019 Benoit LETONDOR
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -14,11 +14,9 @@
  *   limitations under the License.
  */
 
-package com.benoitletondor.mvp.core.presenter.loader;
+package com.benoitletondor.mvp.core.presenter;
 
-import android.support.annotation.NonNull;
-
-import com.benoitletondor.mvp.core.presenter.Presenter;
+import androidx.annotation.NonNull;
 
 /**
  * Factory to create a presenter.
