@@ -1,5 +1,5 @@
 /*
- *   Copyright 2017 Benoit LETONDOR
+ *   Copyright 2019 Benoit LETONDOR
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package com.benoitletondor.mvp.core.view;
 
-import android.support.annotation.UiThread;
+import androidx.annotation.UiThread;
 
 /**
  * A view is in charge of displaying the content to the user. The view is called by the presenter
