@@ -2,10 +2,9 @@ package com.benoitletondor.mvp.core.sample.scene.main.impl;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.TextView;
 
-import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
+import com.benoitletondor.mvp.core.presenter.PresenterFactory;
 import com.benoitletondor.mvp.core.sample.R;
 import com.benoitletondor.mvp.core.sample.scene.dialog.impl.DialogFragment;
 import com.benoitletondor.mvp.core.sample.scene.fragment.base.impl.FragmentActivity;
@@ -15,6 +14,7 @@ import com.benoitletondor.mvp.core.view.impl.BaseMVPActivity;
 
 import javax.inject.Inject;
 
+import androidx.annotation.NonNull;
 import dagger.android.AndroidInjection;
 
 /**
@@ -37,57 +37,38 @@ public final class MainActivity extends BaseMVPActivity<MainPresenter, MainView>
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mCounterTextView = (TextView) findViewById(R.id.activity_main_counter_textview);
+        mCounterTextView = findViewById(R.id.activity_main_counter_textview);
 
-        findViewById(R.id.activity_main_minus_button).setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        findViewById(R.id.activity_main_minus_button).setOnClickListener(view -> {
+            if( mPresenter != null )
             {
-                if( mPresenter != null )
-                {
-                    mPresenter.onMinusButtonClick();
-                }
+                mPresenter.onMinusButtonClick();
             }
         });
 
-        findViewById(R.id.activity_main_plus_button).setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        findViewById(R.id.activity_main_plus_button).setOnClickListener(view -> {
+            if( mPresenter != null )
             {
-                if( mPresenter != null )
-                {
-                    mPresenter.onPlusButtonClick();
-                }
+                mPresenter.onPlusButtonClick();
             }
         });
 
-        findViewById(R.id.activity_main_start_fragment_activity_button).setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        findViewById(R.id.activity_main_start_fragment_activity_button).setOnClickListener(view -> {
+            if( mPresenter != null )
             {
-                if( mPresenter != null )
-                {
-                    mPresenter.onStartFragmentActivityButtonClicked();
-                }
+                mPresenter.onStartFragmentActivityButtonClicked();
             }
         });
 
-        findViewById(R.id.activity_main_start_dialog_fragment_button).setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
+        findViewById(R.id.activity_main_start_dialog_fragment_button).setOnClickListener(view -> {
+            if( mPresenter != null )
             {
-                if( mPresenter != null )
-                {
-                    mPresenter.onShowDialogFragmentButtonClicked();
-                }
+                mPresenter.onShowDialogFragmentButtonClicked();
             }
         });
     }
 
+    @NonNull
     @Override
     protected PresenterFactory<MainPresenter> getPresenterFactory()
     {
