@@ -1,5 +1,5 @@
 /*
- *   Copyright 2017 Benoit LETONDOR
+ *   Copyright 2019 Benoit LETONDOR
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -17,51 +17,39 @@
 package com.benoitletondor.mvp.core.view.impl;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
-import android.support.v4.app.LoaderManager;
-import android.support.v4.content.Loader;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 
 import com.benoitletondor.mvp.core.presenter.Presenter;
-import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
-import com.benoitletondor.mvp.core.presenter.loader.PresenterLoader;
+import com.benoitletondor.mvp.core.presenter.PresenterFactory;
 import com.benoitletondor.mvp.core.view.View;
 
-import java.util.concurrent.atomic.AtomicBoolean;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProviders;
 
-public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> extends AppCompatActivity implements LoaderManager.LoaderCallbacks<P>
+public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> extends AppCompatActivity
 {
-    private final static String TAG = BaseMVPActivity.class.getSimpleName();
-    private final static int LOADER_ID = 17051988;
-
     /**
      * The presenter for this view
      */
-    @SuppressWarnings("WeakerAccess")
     @Nullable
     protected P mPresenter;
     /**
      * Is this the first start of the activity (after onCreate)
      */
     private boolean mFirstStart;
-    /**
-     * Do we need to call {@link #doStart()} from the {@link #onLoadFinished(Loader, P)} method.
-     * Will be true if presenter wasn't loaded when {@link #onStart()} is reached
-     */
-    @VisibleForTesting
-    final AtomicBoolean mNeedToCallStart = new AtomicBoolean(false);
 
 // ------------------------------------------->
 
     @Override
+    @SuppressWarnings("unchecked")
     protected void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
 
         mFirstStart = true;
-        getSupportLoaderManager().initLoader(LOADER_ID, null, this);
+        mPresenter = (P) ViewModelProviders.of(this, new PresenterFactoryWrapper(getPresenterFactory())).get(ViewModel.class);
     }
 
     @Override
@@ -69,12 +57,7 @@ public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> ex
     {
         super.onStart();
 
-        if( mPresenter == null )
-        {
-            mNeedToCallStart.set(true);
-            Log.d(TAG, "Start postponed, presenter not ready");
-        }
-        else
+        if( mPresenter != null )
         {
             doStart();
         }
@@ -91,14 +74,21 @@ public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> ex
         super.onStop();
     }
 
+    @Override
+    protected void onDestroy()
+    {
+        mPresenter = null;
+
+        super.onDestroy();
+    }
+
 // ------------------------------------------->
 
     /**
      * Call the presenter callbacks for onStart
      */
     @SuppressWarnings("unchecked")
-    @VisibleForTesting
-    void doStart()
+    private void doStart()
     {
         assert mPresenter != null;
 
@@ -112,8 +102,7 @@ public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> ex
     /**
      * Call the presenter callbacks for onStop
      */
-    @VisibleForTesting
-    void doStop()
+    private void doStop()
     {
         assert mPresenter != null;
 
@@ -122,34 +111,11 @@ public abstract class BaseMVPActivity<P extends Presenter<V>, V extends View> ex
         mPresenter.onViewDetached();
     }
 
-    @Override
-    public final Loader<P> onCreateLoader(int id, Bundle args)
-    {
-        return new PresenterLoader<>(this, getPresenterFactory());
-    }
-
-    @Override
-    public final void onLoadFinished(Loader<P> loader, P presenter)
-    {
-        mPresenter = presenter;
-
-        if( mNeedToCallStart.compareAndSet(true, false) )
-        {
-            doStart();
-            Log.d(TAG, "Postponed start called");
-        }
-    }
-
-    @Override
-    public final void onLoaderReset(Loader<P> loader)
-    {
-        mPresenter = null;
-    }
-
     /**
      * Get the presenter factory implementation for this view
      *
      * @return the presenter factory
      */
+    @NonNull
     protected abstract PresenterFactory<P> getPresenterFactory();
 }
