@@ -1,5 +1,5 @@
 /*
- *   Copyright 2017 Benoit LETONDOR
+ *   Copyright 2019 Benoit LETONDOR
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -17,60 +17,45 @@
 package com.benoitletondor.mvp.core.view.impl;
 
 import android.app.Dialog;
-import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
-import android.support.v4.app.DialogFragment;
-import android.support.v4.app.LoaderManager;
-import android.support.v4.content.Loader;
-import android.util.Log;
 import android.view.View;
 
 import com.benoitletondor.mvp.core.presenter.Presenter;
-import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
-import com.benoitletondor.mvp.core.presenter.loader.PresenterLoader;
+import com.benoitletondor.mvp.core.presenter.PresenterFactory;
 
-import java.util.concurrent.atomic.AtomicBoolean;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
+import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProviders;
 
 @SuppressWarnings("unused")
-public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends com.benoitletondor.mvp.core.view.View> extends DialogFragment implements LoaderManager.LoaderCallbacks<P>
+public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends com.benoitletondor.mvp.core.view.View> extends DialogFragment
 {
-    private final static String TAG = BaseMVPFragment.class.getName();
-    private final static int LOADER_ID = 17051992;
-
     /**
      * The presenter for this view
      */
-    @SuppressWarnings("WeakerAccess")
     @Nullable
     protected P mPresenter;
     /**
      * Is this the first start of the fragment (after onCreate)
      */
     private boolean mFirstStart;
-    /**
-     * Do we need to call {@link #doStart()} from the {@link #onLoadFinished(Loader, P)} method.
-     * Will be true if presenter wasn't loaded when {@link #onStart()} is reached
-     */
-    @VisibleForTesting
-    final AtomicBoolean mNeedToCallStart = new AtomicBoolean(false);
 
 // ------------------------------------------->
 
     @Override
+    @SuppressWarnings("unchecked")
     public void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
 
         mFirstStart = true;
-
-        getLoaderManager().initLoader(LOADER_ID, null, this);
+        mPresenter = (P) ViewModelProviders.of(this, new PresenterFactoryWrapper(getPresenterFactory())).get(ViewModel.class);
     }
 
     @Override
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState)
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState)
     {
         mFirstStart = true;
 
@@ -91,12 +76,7 @@ public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends co
     {
         super.onStart();
 
-        if( mPresenter == null )
-        {
-            mNeedToCallStart.set(true);
-            Log.d(TAG, "Start postponed, presenter not ready");
-        }
-        else
+        if( mPresenter != null )
         {
             doStart();
         }
@@ -114,15 +94,11 @@ public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends co
     }
 
     @Override
-    public void onDismiss(DialogInterface dialog)
+    public void onDestroy()
     {
-        // Force destroy manually cause it's not called by the framework, not sure why yet
-        if( getActivity() != null ) // If get activity == null, it's just a rotation
-        {
-            getLoaderManager().destroyLoader(LOADER_ID);
-        }
+        mPresenter = null;
 
-        super.onDismiss(dialog);
+        super.onDestroy();
     }
 
 // ------------------------------------------->
@@ -131,8 +107,7 @@ public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends co
      * Call the presenter callbacks for onStart
      */
     @SuppressWarnings("unchecked")
-    @VisibleForTesting
-    void doStart()
+    private void doStart()
     {
         assert mPresenter != null;
 
@@ -146,8 +121,7 @@ public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends co
     /**
      * Call the presenter callbacks for onStop
      */
-    @VisibleForTesting
-    void doStop()
+    private void doStop()
     {
         assert mPresenter != null;
 
@@ -156,34 +130,11 @@ public abstract class BaseMVPDialogFragment<P extends Presenter<V>, V extends co
         mPresenter.onViewDetached();
     }
 
-    @Override
-    public final Loader<P> onCreateLoader(int id, Bundle args)
-    {
-        return new PresenterLoader<>(getContext(), getPresenterFactory());
-    }
-
-    @Override
-    public final void onLoadFinished(Loader<P> loader, P presenter)
-    {
-        mPresenter = presenter;
-
-        if( mNeedToCallStart.compareAndSet(true, false) )
-        {
-            doStart();
-            Log.d(TAG, "Postponed start called");
-        }
-    }
-
-    @Override
-    public final void onLoaderReset(Loader<P> loader)
-    {
-        mPresenter = null;
-    }
-
     /**
      * Get the presenter factory implementation for this view
      *
      * @return the presenter factory
      */
+    @NonNull
     protected abstract PresenterFactory<P> getPresenterFactory();
 }
