@@ -3,7 +3,7 @@ package com.benoitletondor.mvp.core.sample.scene.fragment.base.impl;
 import android.os.Bundle;
 import android.view.MenuItem;
 
-import com.benoitletondor.mvp.core.presenter.loader.PresenterFactory;
+import com.benoitletondor.mvp.core.presenter.PresenterFactory;
 import com.benoitletondor.mvp.core.sample.R;
 import com.benoitletondor.mvp.core.sample.scene.fragment.base.FragmentPresenter;
 import com.benoitletondor.mvp.core.sample.scene.fragment.base.FragmentView;
@@ -12,6 +12,7 @@ import com.benoitletondor.mvp.core.view.impl.BaseMVPActivity;
 
 import javax.inject.Inject;
 
+import androidx.annotation.NonNull;
 import dagger.android.AndroidInjection;
 
 /**
@@ -50,16 +51,16 @@ public final class FragmentActivity extends BaseMVPActivity<FragmentPresenter, F
     {
         int id = item.getItemId();
 
-        switch (id)
+        if ( id == android.R.id.home )
         {
-            case android.R.id.home:
-                finish();
-                return true;
+            finish();
+            return true;
         }
 
         return super.onOptionsItemSelected(item);
     }
 
+    @NonNull
     @Override
     protected PresenterFactory<FragmentPresenter> getPresenterFactory()
     {
